ahh fuck. doesn't work on local documents in file:/// it really sucks that it's so limited. whatever tho i can still fatten muh index OR just not worry about recolorable assets at all... nah it's too cool to pass up, i think i'll see if putting the symbols under the last article tag will work
this may also be limited too tho because i might not be able to call for
these symbols in the iframes that i MIGHT use for the standalone full
chapter releases. just keep these guys easily sepperated for easy
grabbing if i need to keep the header bar exclusive elements up there,
and the page dweller elements down with them.