desktop bookmarks work and banner is being worked on
rounding these numbers manually is straight up stupid as fuck but it's
done. i also saw on reddit that you can apprently also put inline svg in
css so that kills the worry of the potential iframed pages not having
access to any of these symbols. I'll keep the banner and desktop
bookmarks in the main index and get everything else in the css file. the
mail, fury, and svg will be there too in the case i need to refrence
them in bonus content, i think they'ed be useful