adding plot styling & optimizing color picking
i need to add an aspect ratio to .plot and pick a max width to be done,
but i'd also still like to be able to choose manually what character
reads the little easter egg next to it, i just don't know if i should
make them svg's or pngs. I'll just see which is smaller and pick from
there.

the new coloring stuff is just going like .white.text or
.white.background and putting all that info in the classes in the html,
i will need to go back and change the old stuff but i think it's worth
it to not have random different ways in which i change colors for stuff.