added 2 new fonts and progressed row class
firamono is for quotes and francois is the typical p font. i need to
declare a bigger <h1> tag and maybe look into indenting text, atleast
for the blockquote, and maybe have a min width too.

things to keep in mind when i showed donk's server the zip was i am
going to probably need to make a readme and change the index.html to a
different name, the way pixel did it was pretty smart, to just have a
file called "comic".

when no photos are present everything but the current page does the
blank placeholder thing and it kind of pisses me off