just alot of javascript being done
i've been slacking with the commits tho, so far i got primary working
and it updates the index after a reload, this only functions page to
page in firefoxes file:/// but chromium(edge)'s file:/// works fine. i
may just move the pulldown to the main index file, but really thinking
about it what i got here is this working online (comic fury) so no
changes of structure need to happen there, for local files, i still am
going to attempt to use iframes for page changes, and if i do, the theme
will not be updating for any of the content within them. maybe if i only
put the main <main> tag in the iframe that could maybe save it, but then
i'd loose the next page pulling down working properly...

i could give up on iframes but then the file size will be bigger, i
don't think i'm willing to do that for changing themes. this already is
a feature aint nobody really gone be using for a comic aint nobody
really gone be reading, let alone it being a html file.

i think the way forward is to continue with the original plan where the
structure is this:

index.html
	-nav menu
	-banner
	-extramenu <script>
		-(to at least have SOME functionality with changing the
theme)
	-theme.js
	-footer
iframed page
	-<main> (currentpage lastpages nextpage)
	-<article>
	-pulldown.js

this way while you won't be changing the colors of the bonus content, on
the local file:/// on firefox, it will suposidly work on chromium's
file:/// and it WILL function online.

still being able to change the main background color and colors of the
whites at the top is still sorta cool