added actual retail pages
and started work on the standard page layout, i think i'm going to need
to remove "scroll view" and rename "latest page" to last page. scroll
view can just be replaced with support or something idk yet. and also
the comic fury and rss might have to be removed too, and just be left
with the (non-functional) mailing list icon