gave the whereIbe argument a different number
i made it 500 instead of the negative one it was using, idk why it was
using that one, but it obviously wasn't the move. I've yet to try this
on the os and browser i know it's an issue on but I'm confident that
this fixed it.

i also made the nextpage in index go to standard-page.html